+++ repos/ktor/server/src/jvmMain/kotlin/dev/inmo/micro_utils/repos/ktor/server/key_value/KtorStandartWriteKeyValueRepo.kt
@@ -1,20 +1,21 @@
 package dev.inmo.micro_utils.repos.ktor.server.key_value
 
-import dev.inmo.micro_utils.ktor.server.includeWebsocketHandling
-import dev.inmo.micro_utils.ktor.server.uniload
+import dev.inmo.micro_utils.ktor.server.*
 import dev.inmo.micro_utils.repos.WriteStandardKeyValueRepo
 import dev.inmo.micro_utils.repos.ktor.common.key_value.*
 import io.ktor.application.call
 import io.ktor.routing.Route
 import io.ktor.routing.post
 import kotlinx.serialization.KSerializer
-import kotlinx.serialization.builtins.PairSerializer
+import kotlinx.serialization.builtins.*
 
 fun <K, V> Route.configureWriteStandartKeyValueRepoRoutes (
     originalRepo: WriteStandardKeyValueRepo<K, V>,
     keySerializer: KSerializer<K>,
     valueSerializer: KSerializer<V>,
 ) {
+    val keyValueMapSerializer = MapSerializer(keySerializer, valueSerializer)
+    val keysListSerializer = ListSerializer(keySerializer)
     includeWebsocketHandling(
         onNewValueRoute,
         originalRepo.onNewValue,
@@ -28,18 +29,16 @@ fun <K, V> Route.configureWriteStandartKeyValueRepoRoutes (
     )
 
     post(setRoute) {
-        val (key, value) = call.uniload(
-            KeyValuePostObject.serializer(keySerializer, valueSerializer)
+        val toSet = call.uniload(
+            keyValueMapSerializer
         )
 
-        originalRepo.set(key, value)
+        call.unianswer(Unit.serializer(), originalRepo.set(toSet))
     }
 
     post(unsetRoute) {
-        val key = call.uniload(
-            keySerializer
-        )
+        val toUnset = call.uniload(keysListSerializer)
 
-        originalRepo.unset(key)
+        call.unianswer(Unit.serializer(), originalRepo.unset(toUnset))
     }
 }